+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### accept tags: Added, Changed, Removed, Deprecated, Removed, Fixed, Security
 
+## [3.1.0] - 2021-10-12
+### Added
+- add header X-Content-Type-Options with value nosniff;
+- add full_action_name to state;
+- controller redirect support forward query string
+
 ## [3.0.4] - 2021-09-07
 ### Added
 - create CHANGELOG.md